+++ packages/react-async/src/context.ts
@@ -1,4 +1,0 @@
-import * as React from 'react';
-import {Manager, noopManager} from './manager';
-
-export const AsyncContext = React.createContext<Manager>(noopManager);