+++ packages/react-graphql/src/index.ts
@@ -1,3 +1,9 @@
 export {Query} from './Query';
 export {Prefetch, Props as PrefetchProps} from './Prefetch';
 export {createAsyncQueryComponent} from './async';
+export {
+  AsyncQueryComponentType,
+  GraphQLData,
+  GraphQLVariables,
+  QueryProps,
+} from './types';